AE-backend/requirements.txt
asgiref==3.8.1
cachetools==5.3.3
certifi==2024.2.2
chardet==5.2.0
charset-normalizer==3.3.2
coreapi==2.3.3
coreschema==0.0.4
cssselect==1.2.0
cssutils==2.11.0
Django==5.0.6
django-cors-headers==4.3.1
django-extensions==3.2.3
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
idna==3.7
itypes==1.2.0
Jinja2==3.1.4
lxml==5.2.2
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
pillow==10.3.0
premailer==3.10.0
PyJWT==2.8.0
python-dotenv==1.0.1
reportlab==4.2.0
requests==2.32.2
simplejson==3.19.2
sqlparse==0.5.0
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.1
yagmail==0.15.293
